# Multi cloud picking