--- a/src/materials/blur-material.ts
+++ b/src/materials/blur-material.ts
@@ -5,8 +5,6 @@ import { IUniform } from './types';
 
 export interface IBlurMaterialUniforms {
   [name: string]: IUniform<any>;
-  near: IUniform<number>;
-  far: IUniform<number>;
   screenWidth: IUniform<number>;
   screenHeight: IUniform<number>;
   map: IUniform<Texture | null>;
@@ -16,8 +14,6 @@ export class BlurMaterial extends ShaderMaterial {
   vertexShader = require('./shaders/blur.vert');
   fragmentShader = require('./shaders/blur.frag');
   uniforms: IBlurMaterialUniforms = {
-    near: { type: 'f', value: 0 },
-    far: { type: 'f', value: 0 },
     screenWidth: { type: 'f', value: 0 },
     screenHeight: { type: 'f', value: 0 },
     map: { type: 't', value: null },