--- a/src/types.ts
+++ b/src/types.ts
@@ -56,6 +56,7 @@ export interface IPotree {
 export interface PickPoint {
   position?: Vector3;
   normal?: Vector3;
+  pointCloud?: PointCloudOctree;
   [property: string]: any;
 }