This PR adds efficient point picking in multiple point clouds.

Previously, we had to call `pick()` on each cloud separately and merge the pick results back into a single one. The main drawback of this approach is a severe performance hit. We observed around 5 to 10 ms per pick call, almost all the time is spent in `readPixels()`. We can easily have 10 or more point clouds visible at a time, and the picking times just add up, so the frame rate can easily drop below 10 fps or worse.

This PR adds a method `Potree.pick()` that operates on multiple clouds at the same time. It renders them all to the same render target and calls `readPixels()` only once. In my test scenario with 15 point clouds, the time for picking dropped from 65 ms to 7 ms.

There are also a few minor things I changed / fixed in this PR:
* The `pick()` method now accepts `Camera` instead of `PerspectiveCamera`, i.e., it also works for `OrthographicCamera`
* The material properties `useFilterByNormal` and `filterByNormalThreshold` are now copied to the pick material to ensure they are applied during picking
* I dropped the `datasetNormal` field from the pick result (i.e. the untransformed normal) - we don't use it anymore and it can be recovered easily when needed; also it was inconsistent because no such field existed for the coordinate
* The `near` and `far` uniforms were not used in the shaders so I removed them
* If an orthographic camera was combined with attenuated or adaptive size mode, an undefined `fov` value would have been sent to the shader
* An error message is logged to the console if more than 255 nodes take part in the picking as we have only 8 bits minus the value 0 available - this could have happened before, as well, but is more likely to happen with multiple clouds
* The "point index" color type in the visible rendering was probably broken once picking had been used, because in https://github.com/pnext/three-loader/blob/master/src/point-cloud-octree.ts#L407 `node.pcIndex` is set so in https://github.com/pnext/three-loader/blob/master/src/point-cloud-octree.ts#L198 we would no longer use `this.visibleNodes.indexOf(node)` for this node
* Mainly in preparation to the refactoring, I made many methods that didn't access members `static`