@fahadm 2020-04-26
Hey, I've taken a look at this PR and this looks pretty good to me.
One thing before I merge this, can you do a small write up on how to migrate from single point picking to multi point picking and other changes that you made, I would like to add it in the wiki as a migration guide.