@rhuitl 2020-05-19
Hey Fahad, all other changes besides the multi cloud picking should require no changes for our users. The multi-cloud picking only if people actually want to use it.

The result of a point picking, the `PickPoint` interface, has a new member `pointCloud: PointCloudOctree` that is set to the point cloud where the point has been found.

In order to use the multi-cloud picking, simply change from
```ts
pointcloud.pick(...)
```
to the static method of the Potree class and pass in an array of pointclouds:
```ts
Potree.pick([pointcloud1, pointcloud2], ...)
```

It doesn't look like there are any mentions of the original point picking in the Wiki or the examples, so I'm not sure where you would add this.