@fahadm 2020-05-19
Hey, thanks for the heads up, I'll include that in the next release notes. 

I originally didn't notice that we still retain the `PointCloudOctree#pick`, I thought that the static method was replacing it, that's why I thought we need a migration guide.

I have merged the branch to develop, I'll push out a beta version today, and will do the next release at the end of the month. Thanks!